requirements.txt
flask==1.1.2
marshmallow
dynamorm
pyjwt
flask-restplus
